requirements.txt
pandas
openpyxl
apache-airflow[spark,amazon,celery]
celery[redis]
psycopg2